move models to specific packages
