+++ backend/src/main/java/ch/puzzle/okr/models/alignment/AlignmentInterface.java
@@ -1,4 +1,6 @@
-package ch.puzzle.okr.models;
+package ch.puzzle.okr.models.alignment;
+
+import ch.puzzle.okr.models.Objective;
 
 public interface AlignmentInterface<T> {
