+++ backend/src/main/java/ch/puzzle/okr/models/alignment/KeyResultAlignment.java
@@ -1,4 +1,4 @@
-package ch.puzzle.okr.models;
+package ch.puzzle.okr.models.alignment;
 
 import ch.puzzle.okr.models.keyresult.KeyResult;
