+++ backend/src/test/java/ch/puzzle/okr/service/business/AlignmentSelectionBusinessServiceTest.java
@@ -1,7 +1,7 @@
 package ch.puzzle.okr.service.business;
 
-import ch.puzzle.okr.models.AlignmentSelection;
-import ch.puzzle.okr.models.AlignmentSelectionId;
+import ch.puzzle.okr.models.alignment.AlignmentSelection;
+import ch.puzzle.okr.models.alignment.AlignmentSelectionId;
 import ch.puzzle.okr.service.persistence.AlignmentSelectionPersistenceService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
