+++ backend/src/test/java/ch/puzzle/okr/service/business/OverviewBusinessServiceTest.java
@@ -1,8 +1,8 @@
 package ch.puzzle.okr.service.business;
 
-import ch.puzzle.okr.models.Overview;
-import ch.puzzle.okr.models.OverviewId;
 import ch.puzzle.okr.models.Quarter;
+import ch.puzzle.okr.models.overview.Overview;
+import ch.puzzle.okr.models.overview.OverviewId;
 import ch.puzzle.okr.service.persistence.OverviewPersistenceService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
